Fixed Framebuffer Scale Issue [README]
So I found out I could just change the framebuffer texture size & RenderbufferStorage.

Then I encountered another problem which was, Framebuffer Incomplete because a framebuffer can't have a size of 0 which happened when I was in fullscreen and alt + tabbed...

So what I have to do is skip rendering altogether by checking if ClientSize is Zero, then stop the rendering for that moment.